Use CACurrentMediaTime for deltaTime updates

+++ ToyFlightSimulator Shared/Core/Types/GameTime.swift
@@ -8,22 +8,21 @@
 import MetalKit
 
 struct GameTime {
-    // TODO: Change total game time to use unsinged type
-    private static var _totalGameTime: Float = 0.0
-    private static var _deltaTime: Float = 0.0
+    private static var _totalGameTime: Double = 0.0
+    private static var _deltaTime: Double = 0.0
     
-    public static func UpdateTime(_ deltaTime: Float) {
+    public static func UpdateTime(_ deltaTime: Double) {
         self._deltaTime = deltaTime
         self._totalGameTime += deltaTime
     }
 }
 
 extension GameTime {
-    public static var TotalGameTime: Float {
+    public static var TotalGameTime: Double {
         return self._totalGameTime
     }
     
-    public static var DeltaTime: Float {
+    public static var DeltaTime: Double {
         return self._deltaTime
     }
 }
